Split converting from domain to Legacy object
