+++ tests/Unit/DataAccess/LegacyConverters/DomainToLegacyConverterTest.php
@@ -2,25 +2,25 @@
 
 declare( strict_types = 1 );
 
-namespace WMDE\Fundraising\DonationContext\Tests\Unit\DataAccess;
+namespace WMDE\Fundraising\DonationContext\Tests\Unit\DataAccess\LegacyConverters;
 
 use PHPUnit\Framework\TestCase;
-use WMDE\Fundraising\DonationContext\DataAccess\DoctrineDonationRepository;
+use WMDE\Fundraising\DonationContext\DataAccess\LegacyConverters\DomainToLegacyConverter;
 use WMDE\Fundraising\PaymentContext\Domain\Model\BankTransferPayment;
 use WMDE\Fundraising\PaymentContext\Domain\Model\CreditCardPayment;
 use WMDE\Fundraising\PaymentContext\Domain\Model\PaymentMethod;
 use WMDE\Fundraising\PaymentContext\Domain\Model\SofortPayment;
 
 /**
- * @covers \WMDE\Fundraising\DonationContext\DataAccess\DoctrineDonationRepository
+ * @covers \WMDE\Fundraising\DonationContext\DataAccess\LegacyConverters\DomainToLegacyConverter
  */
-class DoctrineDonationRepositoryTest extends TestCase {
+class DomainToLegacyConverterTest extends TestCase {
 
 	/**
 	 * @dataProvider getPaymentMethodsAndTransferCodes
 	 */
 	public function testGetBankTransferCode_identifierIsReturned( string $expectedOutput, PaymentMethod $payment ): void {
-		$this->assertEquals( $expectedOutput, DoctrineDonationRepository::getBankTransferCode( $payment ) );
+		$this->assertEquals( $expectedOutput, DomainToLegacyConverter::getBankTransferCode( $payment ) );
 	}
 
 	public function getPaymentMethodsAndTransferCodes(): array {
